Add setting to disable Computress' tips from config.ini, tips are diabled by default.

+++ src/PlayerManager.cpp
@@ -758,11 +758,11 @@ void PlayerManager::setFirstUseFlags(CNSocket* sock, CNPacketData* data) {
     sP_CL2FE_REQ_PC_FIRST_USE_FLAG_SET* flag = (sP_CL2FE_REQ_PC_FIRST_USE_FLAG_SET*)data->buf;
     Player* plr = getPlayer(sock);
     if (flag->iFlagCode <= 64) {
-        int newflag = 1ULL << (flag->iFlagCode - 1);
+        uint64_t newflag = 1ULL << (flag->iFlagCode - 1);
         plr->iFirstUseFlag1 |= newflag;
     }
     else {
-        int newflag = 1ULL << (flag->iFlagCode - 64 - 1);
+        uint64_t newflag = 1ULL << (flag->iFlagCode - 64 - 1);
         plr->iFirstUseFlag2 |= newflag;
     }
 }
